+++ examples/ff/ff-parser.py
@@ -56,7 +56,7 @@ print 'Running FF parser'
 parser = Parser()
 parser.add_pattern(
     'planner_exit_code', r'^.*run-planner exit code: (.+)$', type=int, file='driver.log')
-parser.add_pattern('evaluations', r'evaluating (\d+) states', required=False)
+parser.add_pattern('evaluations', r'evaluating (\d+) states')
 parser.add_function(error)
 parser.add_function(coverage)
 parser.add_function(get_plan)